Complete app shifted to StateManagement (Provider) and better code and folder structure
