+++ lib/models/user.model.dart
@@ -1,15 +1,4 @@
-// class User {
-//   final int index;
-//   final String about;
-//   final String name;
-//   final String email;
-//   final String picture;
-
-//   User(this.index,this.about,this.name,this.email,this.picture);
-// }
-
-
-String capitalize(String s) => s[0].toUpperCase() + s.substring(1);
+import '../commonFunctions.dart' as CommonFunctions;
 
 class User {
   final int index;
@@ -25,6 +14,6 @@ class User {
   User({this.index,this.fname,this.lname,this.email,this.phone,this.pictureSmall,this.pictureLarge,this.gender,this.dob});
 
   String fullName(){
-    return capitalize(this.fname) + " " + capitalize(this.lname);
+    return CommonFunctions.capitalize(this.fname) + " " + CommonFunctions.capitalize(this.lname);
   }
 }
